use string.gmatch to prevent _HEAL_ABSORB being accounted
string,find passes also passed HEAL_ABSORBED if we search for _HEAL
This is problematic in some cases as the arguments differ, HEAL_ABSORB may be handled seperately in the future. Thus we will use string.gmatch with "_HEAL%$" pattern, which demands that _HEAL is the very last part of the event string.